Removed business logic from SearchScreen.java and placed it into SearchLogic.java. Also added comments on what methods take and return if applicable.
